64-bit support works partially

+++ apa/hdd_fio.c
@@ -610,8 +610,8 @@ int ioctl2DeleteLastSub(hdd_file_slot_t *fileSlot)
 	return rv;
 }
 
-int hddIoctl2(iop_file_t *f, int req, void *argp, unsigned int arglen,
-			  void *bufp, unsigned int buflen)
+int hddIoctl2(iop_file_t *f, int req, void *argp, size_t arglen,
+			  void *bufp, size_t buflen)
 {
 	u32 rv=0;
 	hdd_file_slot_t *fileSlot=f->privdata;
@@ -724,7 +724,7 @@ int devctlSetOsdMBR(u32 device, hddSetOsdMBR_t *mbrInfo)
 }
 
 int hddDevctl(iop_file_t *f, const char *devname, int cmd, void *arg,
-			  unsigned int arglen, void *bufp, unsigned int buflen)
+			  size_t arglen, void *bufp, size_t buflen)
 {
 	int	rv=0;
